refactor: Move WebsiteCache into namespace mediaelch::network
It is not just for scrapers.
