+++ src/scrapers/imdb/ImdbApi.h
@@ -77,7 +77,7 @@ private:
 private:
     const QString m_language;
     mediaelch::network::NetworkManager m_network;
-    WebsiteCache m_cache;
+    network::WebsiteCache m_cache;
 };
 
 } // namespace scraper
