+++ src/scrapers/movie/adultdvdempire/AdultDvdEmpireApi.h
@@ -40,7 +40,7 @@ private:
 
 private:
     mediaelch::network::NetworkManager m_network;
-    WebsiteCache m_cache;
+    network::WebsiteCache m_cache;
 };
 
 } // namespace scraper
