+++ src/scrapers/movie/hotmovies/HotMoviesApi.h
@@ -40,7 +40,7 @@ private:
 
 private:
     mediaelch::network::NetworkManager m_network;
-    WebsiteCache m_cache;
+    network::WebsiteCache m_cache;
 };
 
 } // namespace scraper
