+++ src/scrapers/movie/videobuster/VideoBusterApi.h
@@ -43,7 +43,7 @@ private:
 
 private:
     mediaelch::network::NetworkManager m_network;
-    WebsiteCache m_cache;
+    network::WebsiteCache m_cache;
 };
 
 } // namespace scraper
