+++ src/scrapers/music/AllMusic.h
@@ -32,7 +32,7 @@ public:
 
 private:
     network::NetworkManager m_network;
-    WebsiteCache m_cache;
+    network::WebsiteCache m_cache;
 };
 
 class AllMusic : public QObject
