+++ src/scrapers/music/TheAudioDb.h
@@ -42,7 +42,7 @@ public:
 
 private:
     network::NetworkManager m_network;
-    WebsiteCache m_cache;
+    network::WebsiteCache m_cache;
     QString m_tadbApiKey;
 };
