+++ src/scrapers/tv_show/thetvdb/TheTvDbApi.h
@@ -117,7 +117,7 @@ private:
     const QString m_language;
     mediaelch::network::NetworkManager m_network;
     ApiToken m_token;
-    WebsiteCache m_cache;
+    network::WebsiteCache m_cache;
 };
 
 } // namespace scraper
